add all reports to a sub-folder
