+++ src/runtime_src/core/tools/common/reports/ReportAieMem.h
@@ -5,7 +5,7 @@
 #define __ReportAieMem_h_
 
 // Please keep external include file dependencies to a minimum
-#include "Report.h"
+#include "tools/common/Report.h"
 
 class ReportAieMem : public Report {
  public:
