+++ src/runtime_src/core/tools/common/reports/ReportAieShim.h
@@ -14,20 +14,20 @@
  * under the License.
  */
 
-#ifndef __ReportQspiStatus_h_
-#define __ReportQspiStatus_h_
+#ifndef __ReportAieShim_h_
+#define __ReportAieShim_h_
 
 // Please keep external include file dependencies to a minimum
-#include "Report.h"
+#include "tools/common/Report.h"
 
-class ReportQspiStatus : public Report {
+class ReportAieShim : public Report {
  public:
-  ReportQspiStatus() : Report("qspi-status", "QSPI write protection status", true /*deviceRequired*/) { /*empty*/ };
+  ReportAieShim() : Report("aieshim", "AIE shim tile status", true /*deviceRequired*/) { /*empty*/ };
 
  // Child methods that need to be implemented
  public:
-  virtual void getPropertyTreeInternal(const xrt_core::device * device, boost::property_tree::ptree &pt) const;
-  virtual void getPropertyTree20202(const xrt_core::device * devicee, boost::property_tree::ptree &pt) const;
+  virtual void getPropertyTreeInternal(const xrt_core::device * _pDevice, boost::property_tree::ptree &_pt) const;
+  virtual void getPropertyTree20202(const xrt_core::device * _pDevicee, boost::property_tree::ptree &_pt) const;
   virtual void writeReport(const xrt_core::device* _pDevice, const boost::property_tree::ptree& _pt, const std::vector<std::string>& _elementsFilter, std::ostream & _output) const;
 };
