+++ src/runtime_src/core/tools/common/reports/ReportDynamicRegion.h
@@ -14,22 +14,20 @@
  * under the License.
  */
 
-#ifndef __ReportHost_h_
-#define __ReportHost_h_
+#ifndef __ReportDynamicRegion_h_
+#define __ReportDynamicRegion_h_
 
-// Please keep eternal include file dependencies to a minimum
-#include "Report.h"
+// Please keep external include file dependencies to a minimum
+#include "tools/common/Report.h"
 
-class ReportHost: public Report {
- private:
-  bool m_is_user;
+class ReportDynamicRegion : public Report {
  public:
-  ReportHost(bool is_user = true) : Report("host", "Host information", false /*device required*/) { /*empty*/ m_is_user = is_user;};
+  ReportDynamicRegion() : Report("dynamic-regions", "Information about the xclbin and the compute units", true /*deviceRequired*/) { /*empty*/ };
 
  // Child methods that need to be implemented
  public:
   virtual void getPropertyTreeInternal(const xrt_core::device * _pDevice, boost::property_tree::ptree &_pt) const;
-  virtual void getPropertyTree20202(const xrt_core::device * _pDevice, boost::property_tree::ptree &_pt) const;
+  virtual void getPropertyTree20202(const xrt_core::device * _pDevicee, boost::property_tree::ptree &_pt) const;
   virtual void writeReport(const xrt_core::device* _pDevice, const boost::property_tree::ptree& _pt, const std::vector<std::string>& _elementsFilter, std::ostream & _output) const;
 };
