+++ src/runtime_src/core/tools/common/reports/ReportPlatforms.cpp
