+++ src/runtime_src/core/tools/common/reports/ReportPsKernels.h
@@ -4,7 +4,7 @@
 #define __ReportPsKernels_h_
 
 // Please keep external include file dependencies to a minimum
-#include "Report.h"
+#include "tools/common/Report.h"
 
 class ReportPsKernels : public Report {
  public:
