+++ src/runtime_src/core/tools/common/reports/ReportVmrStatus.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2020-2021 Xilinx, Inc
+ * Copyright (C) 2022 Xilinx, Inc
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may
  * not use this file except in compliance with the License. A copy of the
@@ -14,21 +14,21 @@
  * under the License.
  */
 
-#ifndef __ReportAsyncError_h_
-#define __ReportAsyncError_h_
+#ifndef __ReportVmrStatus_h_
+#define __ReportVmrStatus_h_
 
 // Please keep external include file dependencies to a minimum
-#include "Report.h"
+#include "tools/common/Report.h"
 
-class ReportAsyncError : public Report {
+class ReportVmrStatus : public Report {
  public:
-  ReportAsyncError() : Report("error", "Asyncronus Error present on the device", true /*deviceRequired*/) { /*empty*/ };
+  ReportVmrStatus() : Report("vmr", "Vmr status on the device", true /*deviceRequired*/) { /*empty*/ };
 
  // Child methods that need to be implemented
  public:
-  virtual void getPropertyTreeInternal(const xrt_core::device * _pDevice, boost::property_tree::ptree &_pt) const;
-  virtual void getPropertyTree20202(const xrt_core::device * _pDevicee, boost::property_tree::ptree &_pt) const;
-  virtual void writeReport(const xrt_core::device* _pDevice, const boost::property_tree::ptree& _pt, const std::vector<std::string>& _elementsFilter,std::ostream & _output) const;
+  virtual void getPropertyTreeInternal(const xrt_core::device * dev, boost::property_tree::ptree &pt) const;
+  virtual void getPropertyTree20202(const xrt_core::device * deve, boost::property_tree::ptree &pt) const;
+  virtual void writeReport(const xrt_core::device* _pDevice, const boost::property_tree::ptree& _pt, const std::vector<std::string>& _elementsFilter, std::ostream & _output) const;
 };
 
 #endif
