+++ src/runtime_src/core/tools/xbmgmt2/CMakeLists.txt
@@ -18,7 +18,8 @@ file(GLOB XBMGMT_V2_BASE_FILES
   "../common/ProgressBar.cpp"
   "../common/BusyBar.cpp"
   "../common/Process.cpp"
-  "../common/Report*.cpp"
+  "../common/reports/*.cpp"
+  "../common/Report.cpp"
   "../common/PsKernelUtilities.cpp"
   "../common/Table2D.cpp"
 )
