+++ src/runtime_src/core/tools/xbutil2/CMakeLists.txt
@@ -15,7 +15,8 @@ file(GLOB XBUTIL_V2_BASE_FILES
   "../common/OptionOptions.cpp"
   "../common/XBHelpMenusCore.cpp"
   "../common/XBHelpMenus.cpp"
-  "../common/Report*.cpp"
+  "../common/reports/*.cpp"
+  "../common/Report.cpp"
   "../common/ProgressBar.cpp"
   "../common/BusyBar.cpp"
   "../common/Process.cpp"
