+++ src/runtime_src/core/tools/xbutil2/SubCmdExamine.cpp
@@ -10,26 +10,26 @@
 
 // ---- Reports ------
 #include "tools/common/Report.h"
-#include "tools/common/ReportAie.h"
-#include "tools/common/ReportAieShim.h"
-#include "tools/common/ReportAieMem.h"
-#include "tools/common/ReportAiePartitions.h"
-#include "tools/common/ReportAsyncError.h"
-#include "tools/common/ReportBOStats.h"
-#include "tools/common/ReportCmcStatus.h"
-#include "tools/common/ReportDynamicRegion.h"
-#include "tools/common/ReportDebugIpStatus.h"
-#include "tools/common/ReportElectrical.h"
-#include "tools/common/ReportFirewall.h"
-#include "tools/common/ReportHost.h"
-#include "tools/common/ReportMailbox.h"
-#include "tools/common/ReportMechanical.h"
-#include "tools/common/ReportMemory.h"
-#include "tools/common/ReportPcieInfo.h"
-#include "tools/common/ReportPlatforms.h"
-#include "tools/common/ReportPsKernels.h"
-#include "tools/common/ReportQspiStatus.h"
-#include "tools/common/ReportThermal.h"
+#include "tools/common/reports/ReportAie.h"
+#include "tools/common/reports/ReportAieShim.h"
+#include "tools/common/reports/ReportAieMem.h"
+#include "tools/common/reports/ReportAiePartitions.h"
+#include "tools/common/reports/ReportAsyncError.h"
+#include "tools/common/reports/ReportBOStats.h"
+#include "tools/common/reports/ReportCmcStatus.h"
+#include "tools/common/reports/ReportDynamicRegion.h"
+#include "tools/common/reports/ReportDebugIpStatus.h"
+#include "tools/common/reports/ReportElectrical.h"
+#include "tools/common/reports/ReportFirewall.h"
+#include "tools/common/reports/ReportHost.h"
+#include "tools/common/reports/ReportMailbox.h"
+#include "tools/common/reports/ReportMechanical.h"
+#include "tools/common/reports/ReportMemory.h"
+#include "tools/common/reports/ReportPcieInfo.h"
+#include "tools/common/reports/ReportPlatforms.h"
+#include "tools/common/reports/ReportPsKernels.h"
+#include "tools/common/reports/ReportQspiStatus.h"
+#include "tools/common/reports/ReportThermal.h"
 
 // Note: Please insert the reports in the order to be displayed (alphabetical)
   ReportCollection SubCmdExamineInternal::uniqueReportCollection = {
