+++ src/runtime_src/core/tools/xbutil2/SubCmdValidate.cpp
@@ -15,7 +15,7 @@
 #include "core/tools/common/EscapeCodes.h"
 #include "core/tools/common/Process.h"
 #include "tools/common/Report.h"
-#include "tools/common/ReportPlatforms.h"
+#include "tools/common//reports/ReportPlatforms.h"
 #include "tools/common/XBHelpMenus.h"
 #include "tools/common/XBHelpMenusCore.h"
 #include "tools/common/XBUtilitiesCore.h"
